26-08-2024

+++ src/test/java/testBase/BaseClass.java
@@ -3,6 +3,8 @@ package testBase;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.Date;
@@ -10,55 +12,88 @@ import java.util.Properties;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
 
-import org.apache.logging.log4j.LogManager;//log4j
-import org.apache.logging.log4j.Logger;   //log4j
+
+import org.apache.logging.log4j.LogManager;  //Log4j
+import org.apache.logging.log4j.Logger;  //Log4j
 
 
 public class BaseClass {
 
-	static public WebDriver driver;
-	public Logger logger;
-	public Properties p;
-	
+public static WebDriver driver;
+public Logger logger;  //Log4j
+public Properties p;
 	
 	@BeforeClass(groups= {"sanity","regression","master"})
-	@Parameters({"os", "browser"})
-	public void setup(String os, String br) throws IOException
-	
+	@Parameters({"os","browser"})
+	public void setup(String os, String br) throws IOException, URISyntaxException
 	{
-		//loading properties file
-		 FileReader file=new FileReader(".//src//test//resorces//config.properties");
-		 p=new Properties();
-		 p.load(file);
-		
-		
-		//loading log4j 
-		logger=LogManager.getLogger(this.getClass());//Log4j
+		//Loading config.properties file
+		FileReader file=new FileReader("./src//test//resources//config.properties");
+		p=new Properties();
+		p.load(file);
 				
+		logger=LogManager.getLogger(this.getClass());  //lOG4J2
 				
-		//launching browser based on condition
-		switch(br.toLowerCase())
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote"))
 		{
-		case "chrome": driver=new ChromeDriver(); break;
-		case "edge": driver=new EdgeDriver(); break;
-		default: System.out.println("No matching browser..");
-					return;
+			DesiredCapabilities capabilities=new DesiredCapabilities();
+			
+			//os
+			if(os.equalsIgnoreCase("windows"))
+			{
+				capabilities.setPlatform(Platform.WIN11);
+			}
+			else if (os.equalsIgnoreCase("mac"))
+			{
+				capabilities.setPlatform(Platform.MAC);
+			}
+			else
+			{
+				System.out.println("No matching os");
+				return;
+			}
+			
+			//browser
+			switch(br.toLowerCase())
+			{
+			case "chrome": capabilities.setBrowserName("chrome"); break;
+			case "edge": capabilities.setBrowserName("MicrosoftEdge"); break;
+			default: System.out.println("No matching browser"); return;
+			}
+			
+			driver=new RemoteWebDriver(new URI("http://localhost:4444/wd/hub").toURL(),capabilities);
 		}
 		
+				
+		if(p.getProperty("execution_env").equalsIgnoreCase("local"))
+		{
+
+			switch(br.toLowerCase())
+			{
+			case "chrome" : driver=new ChromeDriver(); break;
+			case "edge" : driver=new EdgeDriver(); break;
+			case "firefox": driver=new FirefoxDriver(); break;
+			default : System.out.println("Invalid browser name.."); return;
+			}
+		}
+		
+			
 		driver.manage().deleteAllCookies();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-		driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(10));
 		
-		
-		driver.get(p.getProperty("appURL1"));
+		driver.get(p.getProperty("appURL2")); // reading url from properties file.
 		driver.manage().window().maximize();
 	}
 	
@@ -68,25 +103,23 @@ public class BaseClass {
 		driver.quit();
 	}
 	
-
 	public String randomeString()
 	{
-		String generatedString=RandomStringUtils.randomAlphabetic(5);
-		return generatedString;
+		String generatedstring=RandomStringUtils.randomAlphabetic(5);
+		return generatedstring;
 	}
 	
 	public String randomeNumber()
 	{
-		String generatedString=RandomStringUtils.randomNumeric(10);
-		return generatedString;
+		String generatednumber=RandomStringUtils.randomNumeric(10);
+		return generatednumber;
 	}
 	
-	public String randomAlphaNumeric()
+	public String randomeAlphaNumberic()
 	{
-		String str=RandomStringUtils.randomAlphabetic(3);
-		String num=RandomStringUtils.randomNumeric(3);
-		
-		return (str+"@"+num);
+		String generatedstring=RandomStringUtils.randomAlphabetic(3);
+		String generatednumber=RandomStringUtils.randomNumeric(3);
+		return (generatedstring+"@"+generatednumber);
 	}
 	
 	public String captureScreen(String tname) throws IOException {
@@ -104,7 +137,6 @@ public class BaseClass {
 		return targetFilePath;
 
 	}
-	
 	
 	
 }
